+++ lib/braintree/transaction/apple_pay_details.rb
@@ -7,9 +7,11 @@ module Braintree
       attr_reader :cardholder_name
       attr_reader :expiration_month
       attr_reader :expiration_year
+      attr_reader :image_url
       attr_reader :last_4
       attr_reader :payment_instrument_name
       attr_reader :source_description
+      attr_reader :token
 
       def initialize(attributes)
         set_instance_variables_from_hash attributes unless attributes.nil?